Restrict header display to signed-in users
The header of the site has been updated to only show when a user is signed in. To accomplish this, a user authentication check was added to the ApplicationController using a before action. This change in the application's layout helps maintain content visibility controls and enhances user experience by showing relevant interface elements.