Exclude 'CODEFI' network from views and tests
This commit updates various views and tests to exclude the 'CODEFI' network. The changes ensure that only relevant networks, specifically 'CRP', are displayed and tested, improving the accuracy and relevance of the application's functionality for the intended context. Additionally, it ensures checks for null values on user segments for better reliability in header views.