Update Gemfile and Gemfile.lock for dependency upgrades and adjustments
This commit updates several gem dependencies in the Gemfile, including upgrading Rails to version 7.2.2.1, Puma to 6.4.2, and Importmap Rails to 2.1.0. Additionally, it modifies the Administrate gem version to 1.0.0.beta3 and adjusts various other gems to their latest compatible versions. The Gemfile.lock is also updated to reflect these changes, ensuring the application uses the most recent and stable versions of its dependencies.